chore: split admin/teacher/school/intervention/best-practices routes out of index.ts
Pure mechanical extraction, no behavior change. Continues the split
started in the batch-a PR.

- routes/admin.ts - POST /api/admin/create, GET /api/admin/coordinators
  (carries forward the #209 State Admin scoping fix unchanged),
  POST /api/admin/revive-teacher, POST /api/admin/restore-school,
  POST /api/reset
- routes/teachers.ts - GET/POST /api/teachers (incl. the
  COORDINATOR_ROLES const used by the registration route)
- routes/schools.ts - GET/POST /api/schools
- routes/interventions.ts - POST/GET /api/interventions*,
  POST /api/interventions/:id/promote
- routes/bestPractices.ts - GET /api/best-practices*

Verified: tsc --noEmit clean; booted the real server against a
scratch MongoDB seed and curled all 5 affected endpoints (all 200).
Specifically re-verified #209's State Admin scoping survived the
extraction byte-for-byte: admin.ap@fln.org's /api/admin/coordinators
still returns only 7 AP-state coordinators, not the full national
list.

Closes #130, closes #122, closes #123, closes #128, closes #125
