+++ backend/src/index.ts
@@ -26,6 +26,11 @@ import { registerTicketRoutes } from './routes/tickets';
 import { registerLogbookRoutes } from './routes/logbook';
 import { registerGeoRoutes } from './routes/geo';
 import { registerClassRoutes } from './routes/classes';
+import { registerAdminRoutes } from './routes/admin';
+import { registerTeacherRoutes } from './routes/teachers';
+import { registerSchoolRoutes } from './routes/schools';
+import { registerInterventionRoutes } from './routes/interventions';
+import { registerBestPracticeRoutes } from './routes/bestPractices';
 import { randomUUID } from 'crypto';
 import fs from 'fs';
 import bcrypt from 'bcrypt';
@@ -60,235 +65,12 @@ registerStatsRoutes(app);
   registerTicketRoutes(app);
   registerLogbookRoutes(app);
 
-  // Admin Creation (by Superadmin)
-  app.post('/api/admin/create', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user || user.role !== UserRole.SUPERADMIN) {
-      return res.status(403).json({ error: 'Forbidden. Superadmin only.' });
-    }
-
-    const { name, email, password, role, stateCode, districtCode, blockCode, schoolId, assignedSchools } = req.body;
-    if (!name || !email || !password || !role) {
-      return res.status(400).json({ error: 'Missing required fields.' });
-    }
-
-    // Verify Password complexity (§3.2 A-3)
-    const hasUppercase = /[A-Z]/.test(password);
-    const hasNumber = /[0-9]/.test(password);
-    const hasSpecial = /[!@#$%^&*(),.?":{}|<>]/.test(password);
-    if (password.length < 8 || !hasUppercase || !hasNumber || !hasSpecial) {
-      return res.status(400).json({ error: 'Password does not meet complexity requirements. Must be >= 8 chars and contain uppercase, digit, and special char.' });
-    }
-
-    const users = await dbStore.getUsers();
-    if (users.some(u => u.email.toLowerCase() === email.toLowerCase())) {
-      return res.status(400).json({ error: 'User with this email already exists.' });
-    }
-
-    const newUser: User = {
-      id: 'u_' + Math.random().toString(36).substr(2, 9),
-      name,
-      email: email.toLowerCase(),
-      role: role as UserRole,
-      stateCode: stateCode ? stateCode.toUpperCase() : undefined,
-      districtCode: districtCode ? districtCode.toUpperCase() : undefined,
-      blockCode: blockCode ? blockCode.toUpperCase() : undefined,
-      schoolId: schoolId || undefined,
-      assignedSchools: assignedSchools || undefined
-    };
-
-    await dbStore.addUser(newUser);
-
-    // Add Log entry
-    await dbStore.addLog({
-      id: 'log_' + Date.now(),
-      timestamp: new Date().toISOString(),
-      schoolId: '',
-      schoolName: 'National Framework',
-      userId: user.id,
-      userEmail: user.email,
-      userRole: user.role,
-      activityType: 'verify',
-      status: 'Success',
-      details: `Superadmin created account: ${name} (${role}) for scope ${stateCode || '*'}/${districtCode || '*'}/${blockCode || '*'}`
-    });
-
-    res.json(newUser);
-  });
-
-  // Coordinator registration: state -> district -> block -> school cascade, then
-  // creating a teacher account scoped to the chosen school.
-  const COORDINATOR_ROLES = [UserRole.SUPERADMIN, UserRole.ADMIN, UserRole.DISTRICT_ADMIN, UserRole.BLOCK_ADMIN];
+  registerAdminRoutes(app);
 
   registerGeoRoutes(app);
 
-  app.get('/api/teachers', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-    if (![UserRole.SCHOOL, UserRole.BLOCK_ADMIN, UserRole.SUPERADMIN].includes(user.role)) {
-      return res.status(403).json({ error: 'Forbidden.' });
-    }
-
-    const [users, schools, classes, students] = await Promise.all([
-      dbStore.getUsers(),
-      dbStore.getSchools(),
-      dbStore.getClasses(),
-      dbStore.getStudents(),
-    ]);
-    const schoolById = new Map(schools.map(s => [s.id, s]));
-
-    let teachers = users.filter(u => u.role === UserRole.TEACHER);
-    if (user.role === UserRole.SCHOOL) {
-      teachers = teachers.filter(t => t.schoolId === user.schoolId);
-    } else if (user.role === UserRole.BLOCK_ADMIN) {
-      teachers = teachers.filter(t => schoolById.get(t.schoolId || '')?.blockCode === user.blockCode);
-    }
-
-    const enriched = teachers.map(t => {
-      const teacherClasses = classes.filter(c => c.teacherId === t.id);
-      const studentsCount = students.filter(s => s.teacherId === t.id).length;
-      return {
-        ...sanitizeUser(t),
-        classes: teacherClasses.map(c => `${c.className} ${c.section}`),
-        studentsCount,
-        status: t.isBanned ? 'Inactive' : 'Active',
-      };
-    });
-
-    res.json(enriched);
-  });
-
-  app.post('/api/teachers', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user || !COORDINATOR_ROLES.includes(user.role)) {
-      return res.status(403).json({ error: 'Forbidden. Coordinator role required.' });
-    }
-
-    const { firstName, lastName, email, phoneNumber, password, school } = req.body;
-    if (!firstName || !lastName || !email || !password || !school) {
-      return res.status(400).json({ error: 'Missing required fields.' });
-    }
-
-    const hasUppercase = /[A-Z]/.test(password);
-    const hasNumber = /[0-9]/.test(password);
-    const hasSpecial = /[!@#$%^&*(),.?":{}|<>]/.test(password);
-    if (password.length < 8 || !hasUppercase || !hasNumber || !hasSpecial) {
-      return res.status(400).json({ error: 'Password does not meet complexity requirements. Must be >= 8 chars and contain uppercase, digit, and special char.' });
-    }
-
-    const schools = await dbStore.getSchools();
-    const targetSchool = schools.find(s => s.id.toLowerCase() === String(school).toLowerCase());
-    if (!targetSchool) return res.status(400).json({ error: 'Unknown school.' });
-
-    const users = await dbStore.getUsers();
-    if (users.some(u => u.email.toLowerCase() === email.toLowerCase())) {
-      return res.status(400).json({ error: 'User with this email already exists.' });
-    }
-
-    const teacherId = 'u_' + Math.random().toString(36).substr(2, 9);
-    const newTeacher: User = {
-      id: teacherId,
-      name: `${firstName} ${lastName}`,
-      email: email.toLowerCase(),
-      role: UserRole.TEACHER,
-      passwordHash: await bcrypt.hash(password, 10),
-      phoneNumber: phoneNumber || undefined,
-      stateCode: targetSchool.stateCode,
-      districtCode: targetSchool.districtCode,
-      blockCode: targetSchool.blockCode,
-      schoolId: targetSchool.id,
-    };
-
-    await dbStore.addUser(newTeacher);
-
-    await dbStore.addLog({
-      id: 'log_' + Date.now(),
-      timestamp: new Date().toISOString(),
-      schoolId: targetSchool.id,
-      schoolName: targetSchool.name,
-      userId: user.id,
-      userEmail: user.email,
-      userRole: user.role,
-      activityType: 'verify',
-      status: 'Success',
-      details: `Coordinator registered teacher: ${newTeacher.name} at ${targetSchool.name}`,
-    });
-
-    res.json({
-      success: true,
-      message: 'Teacher registered successfully.',
-      data: { teacherId, firstName, lastName, email: newTeacher.email },
-    });
-  });
-
-  // Schools
-  app.get('/api/schools', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-    const schools = await dbStore.getSchools();
-    if (user.role === UserRole.SUPERADMIN || user.role === UserRole.ADMIN) {
-      return res.json(schools);
-    }
-    if (user.role === UserRole.SCHOOL || user.role === UserRole.TEACHER) {
-      return res.json(schools.filter(s => s.id === user.schoolId));
-    }
-    if (user.role === UserRole.VOLUNTEER) {
-      return res.json(schools.filter(s => user.assignedSchools?.includes(s.id)));
-    }
-    if (user.role === UserRole.DISTRICT_ADMIN) {
-      return res.json(schools.filter(s => s.districtCode === user.districtCode));
-    }
-    if (user.role === UserRole.BLOCK_ADMIN) {
-      return res.json(schools.filter(s => s.blockCode === user.blockCode));
-    }
-    res.json(schools);
-  });
-
-  app.post('/api/schools', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user || user.role !== UserRole.SUPERADMIN) {
-      return res.status(403).json({ error: 'Forbidden. Superadmin only.' });
-    }
-
-    const { id, name, stateCode, districtCode, blockCode, strength } = req.body;
-    if (!id || !name || !stateCode || !districtCode || !blockCode) {
-      return res.status(400).json({ error: 'Missing required school fields.' });
-    }
-
-    const schools = await dbStore.getSchools();
-    if (schools.some(s => s.id.toLowerCase() === id.toLowerCase())) {
-      return res.status(400).json({ error: 'School ID already exists.' });
-    }
-
-    const newSch: School = {
-      id: id.toLowerCase(),
-      name,
-      stateCode: stateCode.toUpperCase(),
-      districtCode: districtCode.toUpperCase(),
-      blockCode: blockCode.toUpperCase(),
-      strength: strength || 'low',
-      teachersCount: 0,
-      isAccessLocked: false
-    };
-
-    await dbStore.addSchool(newSch);
-
-    // Add Log entry
-    await dbStore.addLog({
-      id: 'log_' + Date.now(),
-      timestamp: new Date().toISOString(),
-      schoolId: newSch.id,
-      schoolName: newSch.name,
-      userId: user.id,
-      userEmail: user.email,
-      userRole: user.role,
-      activityType: 'verify',
-      status: 'Success',
-      details: `Superadmin onboarded a new school: ${newSch.name} (ID: ${newSch.id})`
-    });
-
-    res.json(newSch);
-  });
+  registerTeacherRoutes(app);
+  registerSchoolRoutes(app);
 
   // Classes
   registerClassRoutes(app);
@@ -2677,132 +2459,6 @@ const students = await dbStore.getStudents();
   }
 
   // Get active coordinators/administrators
-  app.get('/api/admin/coordinators', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const users = await dbStore.getUsers();
-    let filtered = users;
-    if (user.role === UserRole.SCHOOL || user.role === UserRole.TEACHER) {
-      filtered = users.filter(u => u.schoolId === user.schoolId);
-    } else if (user.role === UserRole.VOLUNTEER) {
-      filtered = users.filter(u => user.assignedSchools?.includes(u.schoolId || ''));
-    } else if (user.role === UserRole.DISTRICT_ADMIN) {
-      filtered = users.filter(u => u.districtCode === user.districtCode);
-    } else if (user.role === UserRole.BLOCK_ADMIN) {
-      filtered = users.filter(u => u.blockCode === user.blockCode);
-    } else if (user.role === UserRole.ADMIN) {
-      filtered = users.filter(u => u.stateCode === user.stateCode);
-    }
-    res.json(filtered.map(sanitizeUser));
-  });
-
-  // Revive Banned Teacher (§6.5)
-  app.post('/api/admin/revive-teacher', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const allowed = [UserRole.SUPERADMIN, UserRole.ADMIN, UserRole.DISTRICT_ADMIN, UserRole.BLOCK_ADMIN];
-    if (!allowed.includes(user.role)) {
-      return res.status(403).json({ error: 'Forbidden: Insufficient privileges.' });
-    }
-
-    const { teacherId } = req.body;
-    if (!teacherId) return res.status(400).json({ error: 'Teacher ID is required.' });
-
-    const users = await dbStore.getUsers();
-    const teacher = users.find(u => u.id === teacherId && u.role === UserRole.TEACHER);
-    if (!teacher) return res.status(404).json({ error: 'Teacher not found.' });
-
-    await dbStore.updateUser(teacher.id, {
-      delayedAttemptsCount: 0,
-      isBanned: false
-    });
-
-    if (teacher.schoolId) {
-      const schoolTeachers = users.filter(u => u.role === UserRole.TEACHER && u.schoolId === teacher.schoolId);
-      const anyBanned = schoolTeachers.some(t => t.isBanned && t.id !== teacher.id);
-      if (!anyBanned) {
-        await dbStore.updateSchool(teacher.schoolId, { isAccessLocked: false });
-      }
-    }
-
-    await dbStore.addLog({
-      id: 'log_' + Date.now(),
-      timestamp: new Date().toISOString(),
-      schoolId: teacher.schoolId || 'N/A',
-      schoolName: 'GPS',
-      userId: user.id,
-      userEmail: user.email,
-      userRole: user.role,
-      activityType: 'verify',
-      status: 'Success',
-      details: `Revived teacher ${teacher.name} (${teacher.email}) resetting delayed attempts.`
-    });
-
-    res.json({ success: true, message: 'Teacher successfully revived.' });
-  });
-
-  // Restore School Access (§6.5)
-  app.post('/api/admin/restore-school', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const allowed = [UserRole.SUPERADMIN, UserRole.ADMIN];
-    if (!allowed.includes(user.role)) {
-      return res.status(403).json({ error: 'Forbidden: Insufficient privileges.' });
-    }
-
-    const { schoolId } = req.body;
-    if (!schoolId) return res.status(400).json({ error: 'School ID is required.' });
-
-    const schools = await dbStore.getSchools();
-    const school = schools.find(s => s.id === schoolId);
-    if (!school) return res.status(404).json({ error: 'School not found.' });
-
-    await dbStore.updateSchool(schoolId, { isAccessLocked: false });
-
-    // Revive all teachers in this school
-    const users = await dbStore.getUsers();
-    const schoolTeachers = users.filter(u => u.role === UserRole.TEACHER && u.schoolId === schoolId);
-    for (const teacher of schoolTeachers) {
-      await dbStore.updateUser(teacher.id, {
-        delayedAttemptsCount: 0,
-        isBanned: false
-      });
-    }
-
-    await dbStore.addLog({
-      id: 'log_' + Date.now(),
-      timestamp: new Date().toISOString(),
-      schoolId: schoolId,
-      schoolName: school.name,
-      userId: user.id,
-      userEmail: user.email,
-      userRole: user.role,
-      activityType: 'verify',
-      status: 'Success',
-      details: `Manually restored school access for school: ${school.name}`
-    });
-
-    res.json({ success: true, message: 'School access manually restored.' });
-  });
-
-  // ══════════════════════════════════════════
-  // DATABASE RESET (Superadmin only)
-  // ══════════════════════════════════════════
-  // Wipes every collection and re-seeds. Requires an authenticated superadmin.
-  // Deliberately POST-only: a GET reset was removed because it let any prefetch,
-  // crawler, or <img>/link trigger a full database wipe with no credentials.
-  app.post('/api/reset', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user || user.role !== UserRole.SUPERADMIN) {
-      return res.status(403).json({ error: 'Forbidden. Superadmin only.' });
-    }
-    await dbStore.reset();
-    res.json({ success: true, message: 'Database reset to fresh seed data.' });
-  });
-
   // ══════════════════════════════════════════
   // BULK DIAGNOSTIC GENERATION ENDPOINTS
   // ══════════════════════════════════════════
@@ -3142,188 +2798,8 @@ const students = await dbStore.getStudents();
   // --- Intervention Tracking & Best Practices Repository ---
 
   // Create a new intervention
-  app.post('/api/interventions', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user || user.role !== UserRole.TEACHER) {
-      return res.status(403).json({ error: 'Only teachers can record interventions.' });
-    }
-    const { studentId, weakCompetencies, strategyType, strategyDescription, duration, startDate } = req.body;
-    if (!studentId || !weakCompetencies?.length || !strategyType || !strategyDescription) {
-      return res.status(400).json({ error: 'Missing required fields.' });
-    }
-    const students = await dbStore.getStudents();
-    const student = students.find(s => s.id === studentId);
-    if (!student) return res.status(404).json({ error: 'Student not found.' });
-
-    const intervention: Intervention = {
-      id: 'int_' + randomUUID().slice(0, 8),
-      studentId,
-      studentName: student.name,
-      teacherId: user.id,
-      teacherName: user.name,
-      schoolId: user.schoolId || student.schoolId,
-      classId: student.classGroup,
-      className: student.classGroup,
-      section: student.section,
-      weakCompetencies,
-      currentLevel: student.currentLevel,
-      strategyType,
-      strategyDescription,
-      duration: duration || '2 weeks',
-      startDate: startDate || new Date().toISOString().split('T')[0],
-      status: 'active',
-      isPromoted: false,
-      createdAt: new Date().toISOString()
-    };
-    await dbStore.addIntervention(intervention);
-    await dbStore.addLog({
-      id: 'log_' + randomUUID().slice(0, 8),
-      timestamp: new Date().toISOString(),
-      schoolId: user.schoolId || '',
-      schoolName: '',
-      userId: user.id,
-      userEmail: user.email,
-      userRole: user.role,
-      activityType: 'verify',
-      status: 'Success',
-      details: `INTERVENTION: Recorded remedial intervention for ${student.name} — Strategy: ${strategyType}`
-    });
-    res.json(intervention);
-  });
-
-  // List interventions (role-scoped)
-  app.get('/api/interventions', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    let interventions = await dbStore.getInterventions();
-
-    if (user.role === UserRole.TEACHER) {
-      interventions = interventions.filter(i => i.teacherId === user.id);
-    } else if (user.role === UserRole.SCHOOL) {
-      interventions = interventions.filter(i => i.schoolId === user.schoolId);
-    } else if (user.role === UserRole.VOLUNTEER) {
-      const assignedSchools = user.assignedSchools || [];
-      interventions = interventions.filter(i => assignedSchools.includes(i.schoolId));
-    } else if (user.role === UserRole.BLOCK_ADMIN) {
-      const schools = await dbStore.getSchools();
-      const blockSchools = schools.filter(s => s.blockCode === user.blockCode).map(s => s.id);
-      interventions = interventions.filter(i => blockSchools.includes(i.schoolId));
-    }
-    // District Admin, Admin, Superadmin see all
-    res.json(interventions);
-  });
-
-  // Get single intervention
-  app.get('/api/interventions/:id', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-    const interventions = await dbStore.getInterventions();
-    const intervention = interventions.find(i => i.id === req.params.id);
-    if (!intervention) return res.status(404).json({ error: 'Intervention not found.' });
-    res.json(intervention);
-  });
-
-  // Promote intervention to Best Practice (teacher only)
-  app.post('/api/interventions/:id/promote', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user || user.role !== UserRole.TEACHER) {
-      return res.status(403).json({ error: 'Only teachers can promote interventions.' });
-    }
-    const interventions = await dbStore.getInterventions();
-    const intervention = interventions.find(i => i.id === req.params.id);
-    if (!intervention) return res.status(404).json({ error: 'Intervention not found.' });
-    if (intervention.teacherId !== user.id) {
-      return res.status(403).json({ error: 'You can only promote your own interventions.' });
-    }
-    if (!intervention.outcome?.improved) {
-      return res.status(400).json({ error: 'Only interventions with confirmed improvement can be promoted.' });
-    }
-    if (intervention.isPromoted) {
-      return res.status(400).json({ error: 'This intervention is already promoted.' });
-    }
-
-    const bp: BestPractice = {
-      id: 'bp_' + randomUUID().slice(0, 8),
-      interventionId: intervention.id,
-      teacherId: intervention.teacherId,
-      teacherName: intervention.teacherName,
-      schoolId: intervention.schoolId,
-      weakCompetencies: intervention.weakCompetencies,
-      strategyType: intervention.strategyType,
-      strategyDescription: intervention.strategyDescription,
-      levelBefore: intervention.outcome.previousLevel,
-      levelAfter: intervention.outcome.newLevel || intervention.outcome.previousLevel,
-      levelJump: (intervention.outcome.newLevel || 0) - intervention.outcome.previousLevel,
-      duration: intervention.duration,
-      tags: [
-        ...intervention.weakCompetencies,
-        intervention.strategyType.replace('_', ' '),
-        intervention.className
-      ],
-      viewCount: 0,
-      createdAt: new Date().toISOString()
-    };
-
-    await dbStore.addBestPractice(bp);
-    await dbStore.updateIntervention(intervention.id, { isPromoted: true, promotedAt: new Date().toISOString() });
-
-    await dbStore.addLog({
-      id: 'log_' + randomUUID().slice(0, 8),
-      timestamp: new Date().toISOString(),
-      schoolId: user.schoolId || '',
-      schoolName: '',
-      userId: user.id,
-      userEmail: user.email,
-      userRole: user.role,
-      activityType: 'verify',
-      status: 'Success',
-      details: `BEST PRACTICE: Teacher ${user.name} promoted intervention for ${intervention.studentName} to Best Practices Repository`
-    });
-    res.json(bp);
-  });
-
-  // Search/list Best Practices Repository (all roles)
-  app.get('/api/best-practices', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    let bestPractices = await dbStore.getBestPractices();
-    const { search, competency, strategy, sort } = req.query;
-
-    if (search && typeof search === 'string') {
-      const q = search.toLowerCase();
-      bestPractices = bestPractices.filter(bp =>
-        bp.strategyDescription.toLowerCase().includes(q) ||
-        bp.teacherName.toLowerCase().includes(q) ||
-        bp.weakCompetencies.some(c => c.toLowerCase().includes(q)) ||
-        bp.tags.some(t => t.toLowerCase().includes(q))
-      );
-    }
-    if (competency && typeof competency === 'string') {
-      bestPractices = bestPractices.filter(bp => bp.weakCompetencies.includes(competency));
-    }
-    if (strategy && typeof strategy === 'string') {
-      bestPractices = bestPractices.filter(bp => bp.strategyType === strategy);
-    }
-    if (sort === 'level_jump') {
-      bestPractices.sort((a, b) => b.levelJump - a.levelJump);
-    } else {
-      bestPractices.sort((a, b) => new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime());
-    }
-    res.json(bestPractices);
-  });
-
-  // Get single Best Practice (increment view count)
-  app.get('/api/best-practices/:id', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-    const bestPractices = await dbStore.getBestPractices();
-    const bp = bestPractices.find(b => b.id === req.params.id);
-    if (!bp) return res.status(404).json({ error: 'Best practice not found.' });
-    await dbStore.updateBestPractice(bp.id, { viewCount: (bp.viewCount || 0) + 1 });
-    res.json({ ...bp, viewCount: (bp.viewCount || 0) + 1 });
-  });
+  registerInterventionRoutes(app);
+  registerBestPracticeRoutes(app);
 
   // In development, serve the frontend using Vite development middleware.
   // In production, serve the built frontend bundle (frontend/dist).
